requirements.txt
# Core dependencies
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
ta>=0.10.0
python-dotenv>=1.0.0

# Backtesting
backtrader>=1.9.78

# Database
sqlalchemy>=2.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Utilities
requests>=2.28.0
aiohttp>=3.8.0

# Dashboard and Alerts
pyTelegramBotAPI>=4.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
